fix project sorting
